requirements.txt
PyYAML
typer[all]
openapi-generator-cli